fix: resolve build warnings
- MobStatsHandler: use Registries.ENTITY_TYPE.getId() instead of deprecated registryEntry
- AttackDamagePenalty: remove redundant .toDouble() conversion
- BlockInteractionMixin: mixin on Block instead of BlockState, fix method signature

+++ src/main/kotlin/omc/boundbyfate/system/proficiency/penalty/AttackDamagePenalty.kt
@@ -30,7 +30,7 @@ class AttackDamagePenalty(private val multiplier: Float) : PenaltyEffect {
         val modifier = EntityAttributeModifier(
             MODIFIER_UUID,
             MODIFIER_NAME,
-            (multiplier - 1.0).toDouble(),
+            (multiplier - 1.0),
             EntityAttributeModifier.Operation.MULTIPLY_TOTAL
         )
         attribute.addTemporaryModifier(modifier)
